# Use reflect.Type as map keys where possible