@zikaeroh 2020-04-05
An example of what I see in my profile before this PR:

![image](https://user-images.githubusercontent.com/48577114/78466984-d0f70400-76bc-11ea-99b4-d01da0efc5fa.png)

Note the critical path from `handleSession` via `Bind` to the `db` packages's `QueryContext`. Now, after:

![image](https://user-images.githubusercontent.com/48577114/78466997-f97efe00-76bc-11ea-8263-bb050cdef526.png)

No more sqlboiler! All hidden, because it's insignificant. :grin: