@zikaeroh 2020-04-05
I was able to run the benchmarks in the boilbench repo:

```
name                      old time/op    new time/op    delta
BoilDelete/boil-4            657ns ± 1%     667ns ± 1%   +1.53%  (p=0.000 n=10+8)
BoilInsert/boil-4           2.14µs ± 1%    2.15µs ± 1%     ~     (p=0.202 n=8+10)
BoilRawBind/boil-4          9.07µs ± 0%    8.18µs ± 0%   -9.82%  (p=0.000 n=9+10)
BoilSelectAll/boil-4        9.75µs ± 0%    8.87µs ± 1%   -9.01%  (p=0.000 n=10+10)
BoilSelectSubset/boil-4     10.4µs ± 1%     9.5µs ± 1%   -8.39%  (p=0.000 n=9+10)
BoilSelectComplex/boil-4    12.8µs ± 1%    11.9µs ± 0%   -7.19%  (p=0.000 n=10+9)
BoilUpdate/boil-4           1.78µs ± 1%    1.79µs ± 1%   +0.86%  (p=0.000 n=10+9)

name                      old alloc/op   new alloc/op   delta
BoilDelete/boil-4             168B ± 0%      168B ± 0%     ~     (all equal)
BoilInsert/boil-4             968B ± 0%      968B ± 0%     ~     (all equal)
BoilRawBind/boil-4          4.30kB ± 0%    4.06kB ± 0%   -5.78%  (p=0.000 n=10+10)
BoilSelectAll/boil-4        3.18kB ± 0%    2.93kB ± 0%   -7.95%  (p=0.000 n=9+10)
BoilSelectSubset/boil-4     3.32kB ± 0%    3.07kB ± 0%   -7.46%  (p=0.000 n=9+9)
BoilSelectComplex/boil-4    4.23kB ± 0%    3.99kB ± 0%   -5.73%  (p=0.000 n=9+10)
BoilUpdate/boil-4             920B ± 0%      920B ± 0%     ~     (all equal)

name                      old allocs/op  new allocs/op  delta
BoilDelete/boil-4             8.00 ± 0%      8.00 ± 0%     ~     (all equal)
BoilInsert/boil-4             21.0 ± 0%      21.0 ± 0%     ~     (all equal)
BoilRawBind/boil-4            47.0 ± 0%      35.0 ± 0%  -25.53%  (p=0.000 n=10+10)
BoilSelectAll/boil-4          60.0 ± 0%      48.0 ± 0%  -20.00%  (p=0.000 n=10+10)
BoilSelectSubset/boil-4       64.0 ± 0%      52.0 ± 0%  -18.75%  (p=0.000 n=10+10)
BoilSelectComplex/boil-4      90.0 ± 0%      78.0 ± 0%  -13.33%  (p=0.000 n=10+10)
BoilUpdate/boil-4             16.0 ± 0%      16.0 ± 0%     ~     (all equal)
```

And the generated tests do pass:

```
go test -cover ./models      
ok  	github.com/volatiletech/boilbench/models	3.038s	coverage: 71.6% of statements
```