@aarondl 2020-04-19
Truly outstanding work. I apologize for not accepting this sooner. We'll have to update the README benchmarks! Thanks very much for this!