@zikaeroh 2020-04-19
Note that the benchmark hasn't been updated in quite a while; a lot of the snippets don't seem to work like some of Xorm (and Kallax is unfortunately obsolete 🙁). Might end up needing to update the chosen query builders/ORMs/ORM-alikes to get a real view on things, but that's a different item.