@aarondl 2020-04-20
@zikaeroh Very true. We're definitely a bit complacent about it currently. But before we do that we need CI :p